Refactor code; allow protected user to be added to list only if list owner follows them

+++ controllers/lists.controller.js
@@ -1,6 +1,8 @@
 "use strict";
 
 const mongoose = require("mongoose");
+const usersController = require("./users.controller");
+const Follow = require("../models/follow.model");
 const Block = require("../models/block.model");
 const List = require("../models/list.model");
 const ListMember = require("../models/list-member.model");
@@ -17,16 +19,18 @@ const createList = async (req, res, next) => {
 };
 const updateList = async (req, res, next) => {
 	const userId = req.userInfo.userId;
-	const { oldName, newName, includeRepeats, includeReplies } = req.body;
+	const { name, newName, includeRepeats, includeReplies } = req.body;
 	try {
-		const filter = { name: oldName, owner: userId };
+		const filter = { name, owner: userId };
 		if (!(await List.countDocuments(filter))) {
 			res.status(404).send("List not found");
 			return;
 		}
-		if (await List.countDocuments({ name: newName, owner: userId })) {
-			res.status(409).send("You already have another list by that name");
-			return;
+		if (newName && name !== newName) {
+			if (await List.countDocuments({ name: newName, owner: userId })) {
+				res.status(409).send("You already have another list by that name");
+				return;
+			}
 		}
 		const updated = await List.findOneAndUpdate(filter, { name: newName, includeRepeats, includeReplies }, { new: true });
 		res.status(200).json({ updated });
@@ -36,11 +40,21 @@ const updateList = async (req, res, next) => {
 };
 const addMember = async (req, res, next) => {
 	const userId = req.userInfo.userId;
-	const { listName, memberId } = req.body;
+	const { name, handle } = req.body;
 	try {
-		const list = await List.findOne({ name: listName, owner: userId });
+		const list = await List.findOne({ name, owner: userId });
 		if (!list) {
 			res.status(404).send("List not found");
+			return;
+		}
+		const member = await usersController.findActiveUserByHandle(handle);
+		if (!member) {
+			res.status(404).send("User not found");
+			return;
+		}
+		const memberId = member._id;
+		if (member.protected && !(await Follow.findOne({ user: memberId, followedBy: userId }))) {
+			res.status(403).send("You are not allowed to perform this action");
 			return;
 		}
 		if (await Block.countDocuments({ user: userId, blockedBy: memberId })) {
@@ -59,15 +73,21 @@ const addMember = async (req, res, next) => {
 };
 const removeMember = async (req, res, next) => {
 	const userId = req.userInfo.userId;
-	const { listName, memberId } = req.body;
+	const { name, handle } = req.body;
 	try {
-		const list = await List.findOne({ name: listName, owner: userId });
+		const list = await List.findOne({ name, owner: userId });
 		if (!list) {
 			res.status(404).send("List not found");
 			return;
 		}
-		const deleted = ListMember.findOneAndDelete({ list: list._id, user: memberId });
-		res.status(200).json({ deleted });
+		const member = await usersController.findUserByHandle(handle);
+		if (!member) {
+			res.status(404).send("User not found");
+			return;
+		}
+		const memberId = member._id;
+		const removed = await ListMember.findOneAndDelete({ list: list._id, user: memberId });
+		res.status(200).json({ removed });
 	} catch (err) {
 		next(err);
 	}
